fixed id bug
the bug was everytime an item was created it increased the id counter.
I fixed this by finding how many items were in each todo list and
then subtracting the counter respectively. While this led to another
bug which was sometimes the id's were negative, i'm just ignoring it
by setting the prebuilt settings to only 1 item in each list